Correção erro do servidor e Redirecionamento dos links.
